Bug fix for potential zero dived by zero
There is a bug with <this._path.getTotalLength() / alength> on line 92. I have a project where this._path.getTotalLength() is computed as 0 and alength also is computed as 0. Dividing 0 by 0 isn't possible in math. When this happens, it breaks the script from going any further. 

My proposed change resolves this problem. If this._path.getTotalLength() / alength is evaluated as NaN, it sets 'test' to 0, otherwise 'test' becomes the computed division.